MELT: Revise the code convention
