+++ src/helm/benchmark/run_specs/melt_run_specs.py
@@ -31,25 +31,18 @@ from helm.benchmark.metrics.common_metric_specs import (
 
 
 @run_spec_function("melt_question_answering_mlqa")
-def get_question_answering_mlqa_spec(prompt_style: str = "normal") -> RunSpec:
-    assert prompt_style in ["weak", "medium", "normal"], f"Invalid prompt style: {prompt_style}"
+def get_melt_question_answering_mlqa_spec(prompt_style: str = "normal") -> RunSpec:
+    scenario_spec = ScenarioSpec(class_name="helm.benchmark.scenarios.melt_scenarios.MELTQAMLQAScenario")
 
-    scenario_spec = ScenarioSpec(class_name="helm.benchmark.scenarios.melt_scenario.MELTQAScenario", args={
-        "dataset_name": "facebook/mlqa",
-        "subset": "mlqa.vi.vi",
-        "passage_prefix": "Ngữ cảnh: ",
-        "question_prefix": "Câu hỏi: ",
-        "splits": {
-            VALID_SPLIT: "validation",
-            TEST_SPLIT: "test",
-        }
-    })
     if prompt_style == "weak":
         instruction = ""
     elif prompt_style == "medium":
-        instruction = "Hãy trả lời câu hỏi bên dưới bằng tiếng Việt với các thông tin được cung cấp trong phần ngữ cảnh. Nếu trong ngữ cảnh không có đủ thông tin , hãy trả lời \"Tôi không biết\"."
-    else:
+        instruction = 'Hãy trả lời câu hỏi bên dưới bằng tiếng Việt với các thông tin được cung cấp trong phần ngữ cảnh. Nếu trong ngữ cảnh không có đủ thông tin , hãy trả lời "Tôi không biết".'
+    elif prompt_style == "normal":
         instruction = "Bạn là một trợ lý hữu dụng sử dụng tiếng Việt, biết tôn trọng và thành thật. Bạn luôn luôn trả lời các câu hỏi một cách có ích nhiều nhất có thể, nhưng đồng thời phải an toàn. Câu trả lời của bạn không được bao gồm các ngôn từ độc hại, phân biệt chủng tộc, phân biệt giới tính, nguy hiểm, nội dung vi phạm pháp luật. Làm ơn hãy chắc chắn câu trả lời của bạn tự nhiên, tích cực và không thiên vị bất cứ cái gì. Nếu có câu hỏi không hợp lý hoặc không rõ ràng thì hãy giải thích tại sao thay vì trả lời không đúng sự thật. Nếu bạn không biết câu trả lời thì đừng chia sẻ thông tin sai sự thật."
+    else:
+        raise ValueError(f"Invalid prompt style: {prompt_style}")
+
     adapter_spec = get_generation_adapter_spec(instructions=instruction, output_noun="Trả lời", max_tokens=128)
 
     return RunSpec(
@@ -63,25 +56,17 @@ def get_question_answering_mlqa_spec(prompt_style: str = "normal") -> RunSpec:
 
 @run_spec_function("melt_question_answering_xquad")
 def get_question_answering_xquad_spec(prompt_style: str = "normal") -> RunSpec:
-    assert prompt_style in ["weak", "medium", "normal"], f"Invalid prompt style: {prompt_style}"
-
-    scenario_spec = ScenarioSpec(class_name="helm.benchmark.scenarios.melt_scenario.MELTQAScenario", args={
-        "dataset_name": "juletxara/xquad_xtreme",
-        "subset": "vi",
-        "passage_prefix": "Ngữ cảnh: ",
-        "question_prefix": "Câu hỏi: ",
-        "splits": {
-            VALID_SPLIT: "translate_train",
-            TEST_SPLIT: "test",
-        }
-    })
+    scenario_spec = ScenarioSpec(class_name="helm.benchmark.scenarios.melt_scenarios.MELTQAXQuADScenario")
 
     if prompt_style == "weak":
         instruction = ""
     elif prompt_style == "medium":
-        instruction = "Hãy trả lời câu hỏi bên dưới bằng tiếng Việt với các thông tin được cung cấp trong phần ngữ cảnh. Nếu trong ngữ cảnh không có đủ thông tin , hãy trả lời \"Tôi không biết\"."
-    else:
+        instruction = 'Hãy trả lời câu hỏi bên dưới bằng tiếng Việt với các thông tin được cung cấp trong phần ngữ cảnh. Nếu trong ngữ cảnh không có đủ thông tin , hãy trả lời "Tôi không biết".'
+    elif prompt_style == "normal":
         instruction = "Bạn là một trợ lý hữu dụng sử dụng tiếng Việt, biết tôn trọng và thành thật. Bạn luôn luôn trả lời các câu hỏi một cách có ích nhiều nhất có thể, nhưng đồng thời phải an toàn. Câu trả lời của bạn không được bao gồm các ngôn từ độc hại, phân biệt chủng tộc, phân biệt giới tính, nguy hiểm, nội dung vi phạm pháp luật. Làm ơn hãy chắc chắn câu trả lời của bạn tự nhiên, tích cực và không thiên vị bất cứ cái gì. Nếu có câu hỏi không hợp lý hoặc không rõ ràng thì hãy giải thích tại sao thay vì trả lời không đúng sự thật. Nếu bạn không biết câu trả lời thì đừng chia sẻ thông tin sai sự thật."
+    else:
+        raise ValueError(f"Invalid prompt style: {prompt_style}")
+
     adapter_spec = get_generation_adapter_spec(instructions=instruction, output_noun="Trả lời", max_tokens=128)
 
     return RunSpec(
@@ -95,21 +80,12 @@ def get_question_answering_xquad_spec(prompt_style: str = "normal") -> RunSpec:
 
 @run_spec_function("melt_summarization_vietnews")
 def get_summarization_vietnews_spec(prompt_style: str = "normal", temperature: float = 1.0) -> RunSpec:
-    assert prompt_style in ["weak", "medium", "normal"], f"Invalid prompt style: {prompt_style}"
     scenario_spec = ScenarioSpec(
-        class_name="helm.benchmark.scenarios.melt_scenario.MELTSummarizationScenario",
+        class_name="helm.benchmark.scenarios.melt_scenarios.MELTSummarizationVietnewsScenario",
         args={
-            "dataset_name": "Yuhthe/vietnews",
             "train_min_length": 64,
             "train_max_length": 256,
             "doc_max_length": 2048,
-            "article_key": "article",
-            "summary_key": "abstract",
-            "splits": {
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            }
         },
     )
 
@@ -117,8 +93,11 @@ def get_summarization_vietnews_spec(prompt_style: str = "normal", temperature: f
         instruction = ""
     elif prompt_style == "medium":
         instruction = "Nhiệm vụ của bạn là tóm tắt đoạn văn bản sau, đưa ra câu trả lời là bản tóm tắt."
-    else:
+    elif prompt_style == "normal":
         instruction = "Bạn là một trợ lý hữu dụng, biết tôn trọng và thành thật. Bạn luôn luôn trả lời các câu hỏi một cách có ích nhiều nhất có thể, nhưng đồng thời phải an toàn. Câu trả lời của bạn không được bao gồm các ngôn từ độc hại, phân biệt chủng tộc, phân biệt giới tính, nguy hiểm, nội dung vi phạm pháp luật. Nhiệm vụ của bạn là tóm tắt đoạn văn bản nằm trong triple backtick. Bài tóm tắt phải đầy đủ các thông tin quan trọng, ngắn gọn và thu hút người đọc. Ngôn ngữ bạn phải sử dụng để tóm tắt là tiếng Việt."
+    else:
+        raise ValueError(f"Invalid prompt style: {prompt_style}")
+
     adapter_spec = get_generation_adapter_spec(
         instructions=instruction,
         input_noun="Đoạn văn",
@@ -139,21 +118,12 @@ def get_summarization_vietnews_spec(prompt_style: str = "normal", temperature: f
 
 @run_spec_function("melt_summarization_wikilingua")
 def get_wsummarization_wikilingua_spec(prompt_style: str = "normal", temperature: float = 1.0) -> RunSpec:
-    assert prompt_style in ["weak", "medium", "normal"], f"Invalid prompt style: {prompt_style}"
     scenario_spec = ScenarioSpec(
-        class_name="helm.benchmark.scenarios.melt_scenario.MELTSummarizationScenario",
+        class_name="helm.benchmark.scenarios.melt_scenarios.MELTSummarizationWikilinguaScenario",
         args={
-            "dataset_name": "GEM/wiki_lingua",
             "train_min_length": 64,
             "train_max_length": 256,
             "doc_max_length": 2048,
-            "article_key": "source",
-            "summary_key": "target",
-            "splits": {
-                TRAIN_SPLIT: "train",
-                VALID_SPLIT: "validation",
-                TEST_SPLIT: "test",
-            }
         },
     )
 
@@ -161,8 +131,11 @@ def get_wsummarization_wikilingua_spec(prompt_style: str = "normal", temperature
         instruction = ""
     elif prompt_style == "medium":
         instruction = "Nhiệm vụ của bạn là tóm tắt đoạn văn bản sau, đưa ra câu trả lời là bản tóm tắt."
-    else:
+    elif prompt_style == "normal":
         instruction = "Bạn là một trợ lý hữu dụng, biết tôn trọng và thành thật. Bạn luôn luôn trả lời các câu hỏi một cách có ích nhiều nhất có thể, nhưng đồng thời phải an toàn. Câu trả lời của bạn không được bao gồm các ngôn từ độc hại, phân biệt chủng tộc, phân biệt giới tính, nguy hiểm, nội dung vi phạm pháp luật. Nhiệm vụ của bạn là tóm tắt đoạn văn bản nằm trong triple backtick. Bài tóm tắt phải đầy đủ các thông tin quan trọng, ngắn gọn và thu hút người đọc. Ngôn ngữ bạn phải sử dụng để tóm tắt là tiếng Việt."
+    else:
+        raise ValueError(f"Invalid prompt style: {prompt_style}")
+
     adapter_spec = get_generation_adapter_spec(
         instructions=instruction,
         input_noun="Đoạn văn",
@@ -179,88 +152,3 @@ def get_wsummarization_wikilingua_spec(prompt_style: str = "normal", temperature
         + get_generative_harms_metric_specs(),
         groups=["melt", "summarization_wikilingua"],
     )
-
-
-@run_spec_function("melt_sentiment_analysis_uitvsfc")
-def get_sentiment_analysis_uitvsfc_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_sentiment_analysis_vlsp2016")
-def get_sentiment_analysis_vlsp2016_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_text_classification_phoatis")
-def get_text_classification_phoatis_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_text_classification_uitvsmec")
-def get_text_classification_uitvsmec_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_knowledge_uitvimmrc")
-def get_knowledge_uitvimmrc_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_knowledge_zaloe2e")
-def get_knowledge_zaloe2e_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_toxicity_detection_uitvictsd")
-def get_toxicity_detection_uitvictsd_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_toxicity_detection_uitvihsd")
-def get_toxicity_detection_uitvihsd_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_information_retrieval_mmarco")
-def get_information_retrieval_mmarco_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_information_retrieval_mrobust04")
-def get_information_retrieval_mrobust04_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_language_modeling_mlqa")
-def get_language_modeling_mlqa_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_language_modeling_vsec")
-def get_language_modeling_vsec_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_reasoning_math")
-def get_math_reasoning_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_reasoning_synthetic_reasoning")
-def get_reasoning_synthetic_reasoning_spec(mode: str) -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_reasoning_synthetic_reasoning_natural")
-def get_reasoning_synthetic_reasoning_natural_spec(difficulty: str) -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_translation_opus100")
-def get_translation_opus100_spec() -> RunSpec:
-    raise NotImplementedError()
-
-
-@run_spec_function("melt_translation_phomt")
-def get_translation_phomt_spec() -> RunSpec:
-    raise NotImplementedError()
